requirements.txt
pyusb
click
coloredlogs
inquirer3>=0.1.0
